# Cache entry point lookups