--- a/.github/workflows/rabbitmq.yml
+++ b/.github/workflows/rabbitmq.yml
@@ -48,10 +48,10 @@ jobs:
         steps:
         -   uses: actions/checkout@v2
 
-        -   name: Set up Python 3.10
-            uses: actions/setup-python@v4
+        -   name: Set up Python 3.9
+            uses: actions/setup-python@v2
             with:
-                python-version: '3.10'
+                python-version: '3.9'
 
         -   name: Install system dependencies
             run: sudo apt update && sudo apt install postgresql