--- a/aiida/manage/tests/pytest_fixtures.py
+++ b/aiida/manage/tests/pytest_fixtures.py
@@ -33,6 +33,7 @@ import typing as t
 import uuid
 import warnings
 
+from importlib_metadata import EntryPoints
 import plumpy
 import pytest
 import wrapt
@@ -759,9 +760,16 @@ def suppress_deprecations(wrapped, _, args, kwargs):
 class EntryPointManager:
     """Manager to temporarily add or remove entry points."""
 
-    @staticmethod
-    def eps():
-        return plugins.entry_point.eps()
+    def __init__(self, entry_points: EntryPoints):
+        self.entry_points = entry_points
+
+    def eps(self) -> EntryPoints:
+        return self.entry_points
+
+    def eps_select(self, group, name=None) -> EntryPoints:
+        if name is None:
+            return self.eps().select(group=group)
+        return self.eps().select(group=group, name=name)
 
     @staticmethod
     def _validate_entry_point(entry_point_string: str | None, group: str | None, name: str | None) -> tuple[str, str]:
@@ -791,7 +799,6 @@ class EntryPointManager:
 
         return group, name
 
-    @suppress_deprecations
     def add(
         self,
         value: type | str,
@@ -817,9 +824,8 @@ class EntryPointManager:
 
         group, name = self._validate_entry_point(entry_point_string, group, name)
         entry_point = plugins.entry_point.EntryPoint(name, value, group)
-        self.eps()[group].append(entry_point)
+        self.entry_points = EntryPoints(self.entry_points + (entry_point,))
 
-    @suppress_deprecations
     def remove(
         self, entry_point_string: str | None = None, *, name: str | None = None, group: str | None = None
     ) -> None:
@@ -835,32 +841,23 @@ class EntryPointManager:
         :raises ValueError: If `entry_point_string` is not a complete entry point string with group and name.
         """
         group, name = self._validate_entry_point(entry_point_string, group, name)
-
-        for entry_point in self.eps()[group]:
-            if entry_point.name == name:
-                self.eps()[group].remove(entry_point)
-                break
-        else:
+        try:
+            self.entry_points[name]
+        except KeyError:
             raise KeyError(f'entry point `{name}` does not exist in group `{group}`.')
+        self.entry_points = EntryPoints((ep for ep in self.entry_points if not (ep.name == name and ep.group == group)))
 
 
 @pytest.fixture
 def entry_points(monkeypatch) -> EntryPointManager:
     """Return an instance of the ``EntryPointManager`` which allows to temporarily add or remove entry points.
 
-    This fixture creates a deep copy of the entry point cache returned by the :func:`aiida.plugins.entry_point.eps`
-    method and then monkey patches that function to return the deepcopy. This ensures that the changes on the entry
-    point cache performed during the test through the manager are undone at the end of the function scope.
-
-    .. note:: This fixture does not use the ``suppress_deprecations`` decorator on purpose, but instead adds it manually
-        inside the fixture's body. The reason is that otherwise all deprecations would be suppressed for the entire
-        scope of the fixture, including those raised by the code run in the test using the fixture, which is not
-        desirable.
-
+    This fixture monkey patches the entry point caches returned by the :func:`aiida.plugins.entry_point.eps` and
+    :func:`aiida.plugins.entry_point.eps_select` methods to a class methods of the EntryPointManager
+    so that we can dynamically add / remove entry points.
+    Note that we do not need a deepcopy here is eps() returns an immutable EntryPoints tuple type.
     """
-    with warnings.catch_warnings():
-        warnings.filterwarnings('ignore', category=DeprecationWarning)
-        eps_copy = copy.deepcopy(plugins.entry_point.eps())
-    monkeypatch.setattr(plugins.entry_point, 'eps', lambda: eps_copy)
-    plugins.entry_point.eps_select.cache_clear()
-    yield EntryPointManager()
+    epm = EntryPointManager(plugins.entry_point.eps())
+    monkeypatch.setattr(plugins.entry_point, 'eps', epm.eps)
+    monkeypatch.setattr(plugins.entry_point, 'eps_select', epm.eps_select)
+    yield epm