--- a/aiida/manage/tests/pytest_fixtures.py
+++ b/aiida/manage/tests/pytest_fixtures.py
@@ -21,7 +21,6 @@ from __future__ import annotations
 
 import asyncio
 import contextlib
-import copy
 import inspect
 import io
 import os