--- a/aiida/plugins/entry_point.py
+++ b/aiida/plugins/entry_point.py
@@ -332,14 +332,13 @@ def get_entry_point_from_class(class_module: str, class_name: str) -> Tuple[Opti
     :param class_name: name of the class
     :return: a tuple of the corresponding group and entry point or None if not found
     """
-    for group in get_entry_point_groups():
-        for entry_point in get_entry_points(group):
+    for entry_point in eps():
 
-            if entry_point.module != class_module:
-                continue
+        if entry_point.module != class_module:
+            continue
 
-            if entry_point.attr == class_name:
-                return group, entry_point
+        if entry_point.attr == class_name:
+            return entry_point.group, entry_point
 
     return None, None