--- a/environment.yml
+++ b/environment.yml
@@ -20,7 +20,7 @@ dependencies:
 - jinja2~=3.0
 - jsonschema~=3.0
 - kiwipy[rmq]~=0.7.7
-- importlib-metadata~=5.2
+- importlib-metadata~=6.0
 - numpy~=1.21
 - paramiko>=2.7.2,~=2.7
 - plumpy~=0.21.6