--- a/pyproject.toml
+++ b/pyproject.toml
@@ -38,7 +38,7 @@ dependencies = [
     "jinja2~=3.0",
     "jsonschema~=3.0",
     "kiwipy[rmq]~=0.7.7",
-    "importlib-metadata~=4.13",
+    "importlib-metadata~=5.2",
     "numpy~=1.21",
     "paramiko~=2.7,>=2.7.2",
     "plumpy~=0.21.6",