--- a/requirements/requirements-py-3.11.txt
+++ b/requirements/requirements-py-3.11.txt
@@ -56,7 +56,7 @@ graphviz==0.20.1
 greenlet==2.0.2
 idna==3.4
 imagesize==1.4.1
-importlib-metadata==5.2.0
+importlib-metadata==6.8.0
 iniconfig==2.0.0
 ipykernel==6.23.2
 ipython==8.14.0