In #6091 we have sped up `aiida.orm` import significantly by delaying the call to `get_type_string_from_class`. However, we did not solve the underlying issue -- the reason why this call is so expensive is because it essentially involves a quadruply-nested for loop. 

Here I cache the results of `importlib_metadata.entry_points.select()` calls, which should get rid of the two inner loops coming from `importlib_metadata` code. This change shaves off around 25ms from `verdi` invocation, but should help throughout the code base. I am happy to do further benchmarking if needed (what should I benchmark?)

cc @sphuber