@danielhollas 2023-10-02
I looked a bit more carefully at the implementation of the `get_entry_point_from_class`.
In this function, we essentially need to iterate over all existing entry points (in my environment, there are 360 of them!). 
And we're doing this iteration very inefficiently, because we first grab the groups, and then filter entry points in each group. But this filtering is done using `EntryPoints.select()` which itself iterates over all the entry points.

Instead, we can simply iterate over all the entry points directly, see the last commit.

Here are some timings, using a non-existing class to measure the worst-case scenario:
(note that I have disabled cache for `get_entry_point_from_class` for these timings)

### Main branch

```python
In [1]: from aiida.plugins.entry_point import get_entry_point_from_class, eps
In [2]: p = eps() #warmup cache
In [3]: %timeit get_entry_point_from_class('invalid', 'invalid')
21.4 ms ± 108 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

### Caching eps.select() on this branch

```python
In [1]: from aiida.plugins.entry_point import get_entry_point_from_class, eps
In [2]: p = eps(); get_entry_point_from_class('invalid', 'invalid') # warmup caches
In [12]: %timeit  get_entry_point_from_class('invalid', 'invalid')
329 µs ± 2.78 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

### Iterating over all entry points directly

```python
In [1]: from aiida.plugins.entry_point import get_entry_point_from_class, eps
In [2]: p = eps()
In [5]: %timeit get_entry_point_from_class('invalid', 'invalid')
288 µs ± 1.33 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

So the last approach clearly wins, and perhaps we could even remove the cache on top of this function.

I'll do a bit more investigation to see if it still makes sense to cache `eps.select() ` since it is called from other functions as well.