@sphuber 2023-10-02
> EDIT: The last commit does not work, need to investigate more..

The problem is that `get_entry_points` returns a list of `EntryPoint` whereas the `importlib_metadata.entry_points` function returns a list of strings. So you would have to load the entry point for each string in order to be able to get the module and class name. Not sure what the impact of having to load each entry point is going to be for the performance and whether the implementation will then still be faster than simply using `get_entry_points`.