@sphuber 2023-10-06
> EDIT: Updating 5.2.0 would require at the very least updating the EntryPointManager pytest fixture.

This should be straightforward. The `eps` has just changed from a nested mapping to a flat list, so instead of having `add_entry_point` treat it like a dict, we should simply append the entry point to the list. @danielhollas want me to take resolve this so we can wrap up this PR?