@danielhollas 2023-10-06
@sphuber So I should have clarified my hesitation here: I was not sure if you'd be okay with upgrading `importlib_metadata` because I wasn't sure if downstream packages might depend on this package as well. If that is not a concern I'll go ahead and finish this today or over the weekend.

> The eps has just changed from a nested mapping to a flat list, so instead of having add_entry_point treat it like a dict, we should simply append the entry point to the list. @danielhollas want me to take resolve this so we can wrap up this PR?

Indeed, but I stumbled at first because `EntryPoints` is in fact a subclass of `tuple` and not a `list` hence we'll probably need to recreate it instead of modifying it directly.