@sphuber 2023-10-06
No, I think it is fine for use to update the minimum requirement of `importlib_metadat`. Please feel free to go ahead with this